refactor: 重构项目基本结构，引入vcpkg
