+++ include/impl/maria/MariaDriverInstanceImpl.h
@@ -1,19 +1,20 @@
 #pragma once
 #include <sese/db/DriverInstance.h>
-#include <sese/db/impl/SqlitePreparedStatementImpl.h>
+#include <maria/MariaPreparedStatementImpl.h>
 
 namespace sese::db::impl {
 
-    class SESE_DB_API SqliteDriverInstanceImpl : public DriverInstance {
+    class SESE_DB_API MariaDriverInstanceImpl : public DriverInstance {
     public:
-        explicit SqliteDriverInstanceImpl(sqlite3 *conn) noexcept;
-        ~SqliteDriverInstanceImpl() noexcept override;
+        explicit MariaDriverInstanceImpl(MYSQL *conn) noexcept;
+        ~MariaDriverInstanceImpl() noexcept override;
+
         ResultSet::Ptr executeQuery(const char *sql) const noexcept override;
         int64_t executeUpdate(const char *sql) const noexcept override;
         PreparedStatement::Ptr createStatement(const char *sql) const noexcept override;
 
     protected:
-        sqlite3 *conn = nullptr;
+        MYSQL *conn;
     };
 
 }// namespace sese::db::impl
