+++ include/impl/sqlite/SqliteDriverInstanceImpl.h
@@ -1,20 +1,19 @@
 #pragma once
 #include <sese/db/DriverInstance.h>
-#include <sese/db/impl/MariaPreparedStatementImpl.h>
+#include <sqlite/SqlitePreparedStatementImpl.h>
 
 namespace sese::db::impl {
 
-    class SESE_DB_API MariaDriverInstanceImpl : public DriverInstance {
+    class SESE_DB_API SqliteDriverInstanceImpl : public DriverInstance {
     public:
-        explicit MariaDriverInstanceImpl(MYSQL *conn) noexcept;
-        ~MariaDriverInstanceImpl() noexcept override;
-
+        explicit SqliteDriverInstanceImpl(sqlite3 *conn) noexcept;
+        ~SqliteDriverInstanceImpl() noexcept override;
         ResultSet::Ptr executeQuery(const char *sql) const noexcept override;
         int64_t executeUpdate(const char *sql) const noexcept override;
         PreparedStatement::Ptr createStatement(const char *sql) const noexcept override;
 
     protected:
-        MYSQL *conn;
+        sqlite3 *conn = nullptr;
     };
 
 }// namespace sese::db::impl
