+++ include/impl/sqlite/SqliteResultSetImpl.h
